requirements.txt
torch~=2.6.0
transformers~=4.51.2
h11~=0.14.0
pip~=23.2.1
attrs~=23.2.0
wheel~=0.43.0
PySocks~=1.7.1
protobuf~=4.25.6
cryptography~=44.0.0
pillow~=10.3.0
tornado~=6.4
Jinja2~=3.1.4
requests~=2.32.3
colorama~=0.4.6
ipython~=8.24.0
notebook~=7.3.2
PyYAML~=6.0.2
typing_extensions~=4.12.2
sniffio~=1.3.1
trio~=0.26.0
outcome~=1.3.0.post0
click~=8.1.7
Pygments~=2.18.0
httpcore~=1.0.7
idna~=3.7
httpx~=0.28.1
numpy~=1.26.4
charset-normalizer~=3.3.2
setuptools~=71.1.0
pytz~=2024.1
overrides~=7.7.0
cffi~=1.16.0
pyinstaller~=6.9.0
regex~=2024.11.6
sympy~=1.13.1
mpmath~=1.3.0
decorator~=4.4.2
matplotlib~=3.8.4
tqdm~=4.66.4
scipy~=1.13.0
psutil~=5.9.8
packaging~=24.0
opt_einsum~=3.4.0
filelock~=3.15.3
urllib3~=2.2.2
fsspec~=2024.6.0
MarkupSafe~=2.1.5
anyio~=4.7.0
pandas~=2.2.2
sounddevice~=0.5.1
pywin32~=306
certifi~=2024.6.2
networkx~=3.3
safetensors~=0.5.3
boto3~=1.35.97
huggingface-hub~=0.30.2
platformdirs~=4.2.1
tokenizers~=0.21.1
jax~=0.5.2
sentencepiece~=0.2.0
beautifulsoup4~=4.12.3
tensorboardX~=2.6.2.2
kafka-python~=2.1.5
dotenv~=0.9.9
python-dotenv~=1.1.0